+++ backend/GeObras/src/main/java/com/example/geobras/dto/OrcamentoResponseDTO.java
@@ -2,7 +2,7 @@ package com.example.geobras.dto;
 
 public record OrcamentoResponseDTO(
         Double deslocamento,
-        Integer maoDeObra,
+        Double maoDeObra,
         String servico,
         Double aluguelDeEquipamento,
         Double imposto,
